+++ test/requirements.txt
@@ -6,3 +6,4 @@ pytest-timeout==1.3.4
 pytest-xdist==1.31.0
 boto3
 retrying
+gitpython
