+++ test/sagemaker_tests/pytorch/training/conftest.py
@@ -12,19 +12,21 @@
 # language governing permissions and limitations under the License.
 from __future__ import absolute_import
 
-import boto3
 import os
 import logging
 import platform
-import pytest
 import shutil
 import sys
 import tempfile
+
+import pytest
+import boto3
 
 from sagemaker import LocalSession, Session
 from sagemaker.pytorch import PyTorch
 
 from .utils import image_utils
+from test.test_utils import test_reporting
 
 logger = logging.getLogger(__name__)
 logging.getLogger('boto').setLevel(logging.INFO)
@@ -55,6 +57,13 @@ def pytest_addoption(parser):
     parser.addoption('--processor', choices=['gpu', 'cpu'], default='cpu')
     # If not specified, will default to {framework-version}-{processor}-py{py-version}
     parser.addoption('--tag', default=None)
+    parser.addoption('--generate-coverage-doc', default=False, action='store_true',
+                     help='use this option to generate test coverage doc')
+
+
+def pytest_collection_modifyitems(session, config, items):
+    if config.getoption("--generate-coverage-doc"):
+        test_reporting.generate_coverage_doc(items, sagemaker=True, framework="pytorch", job_type="training")
 
 
 @pytest.fixture(scope='session', name='docker_base_name')
