+++ test/test_utils/__init__.py
@@ -1,4 +1,3 @@
-import datetime
 import json
 import os
 import re
@@ -51,9 +50,6 @@ SAGEMAKER_LOCAL_TEST_TYPE = "local"
 SAGEMAKER_REMOTE_TEST_TYPE = "sagemaker"
 
 PUBLIC_DLC_REGISTRY = "763104351884"
-
-# Test coverage file name
-TEST_COVERAGE_FILE = f"test_coverage_report-{datetime.datetime.now().strftime('%Y-%m-%d')}.csv"
 
 
 def is_tf1(image_uri):
